requirements.txt
# pip3 install torch torchvision torchaudio --extra-index-url https://download.pytorch.org/whl/cpu

flask
pinecone-client
sentence-transformers